Create object.go
